LPS-106167 Inline BaseUploadHandler to TestUploadHandler
